requirements.txt
slack-bolt==1.21.3
anthropic==0.49.0
requests==2.32.3
python-dotenv==1.0.1
